Moved window functions to funcs
